+++ src/store/index.js
@@ -5,7 +5,7 @@ import user from './modules/user'
 
 Vue.use(Vuex);
 
-export default new Vuex.Store({
+const store = new Vuex.Store({
   plugins: [createPersistedState({
     storage: window.sessionStorage
   })],
@@ -18,4 +18,6 @@ export default new Vuex.Store({
   modules: {
     user
   }
-})
+});
+
+export default store
