+++ src/views/HuskyAside.vue
@@ -4,7 +4,7 @@
       <el-button type="primary" @click="isCollapse = true" v-if="!isCollapse" class="logo">助手小哈</el-button>
       <el-button type="primary" @click="isCollapse = false" v-if="isCollapse" class="el-icon-s-operation logo"></el-button>
     </div>
-    <el-menu default-active="home" class="husky-nav-menu" @open="handleOpen" @close="handleClose"
+    <el-menu default-active="home" class="husky-nav-menu"
              :collapse="isCollapse" router="router" v-if="isLogin">
       <el-menu-item index="home">
         <i class="el-icon-setting"></i>
@@ -18,12 +18,12 @@
         <el-menu-item index="account_book">账簿</el-menu-item>
         <el-menu-item index="salary" disabled>薪酬</el-menu-item>
       </el-submenu>
-      <el-menu-item index="setting">
+      <el-menu-item index="setting" disabled>
         <i class="el-icon-setting"></i>
         <span slot="title">设置</span>
       </el-menu-item>
     </el-menu>
-    <el-menu default-active="home" class="husky-nav-menu" @open="handleOpen" @close="handleClose"
+    <el-menu default-active="home" class="husky-nav-menu"
              :collapse="isCollapse" router="router" v-if="!isLogin">
       <el-menu-item index="sign_in">
         <i class="el-icon-position"></i>
@@ -45,14 +45,6 @@ export default {
   data: () => {
     return {
       isCollapse: false
-    }
-  },
-  methods: {
-    handleOpen(key, keyPath) {
-      console.log(key, keyPath);
-    },
-    handleClose(key, keyPath) {
-      console.log(key, keyPath);
     }
   },
   computed: {
